refactor: move email templates from utils to services
